NotePad App Server Part
The server application of the program is run on the server. Before running the program, you need to edit the ports, ip and path according to your own server. The server application must be run before running the client.  In this way, you can run the client anywhere as long as you access the server."